Fix references to external dependencies + solution fixes
Add property sheet SolutionDependencies.props which defines user macros that point to the wxWidgets library to use.  The actual property sheet is not committed so it can be locally modified.  Instead, SolutionDependencies.props.default is added, and must be copied to SolutionDependencies.props manually.

Fix the .vcxproj so it uses the property sheet macros to refer to the wxWidgets components instead of hard-coding library paths and .lib names.

Fix the Debug/Win32 target - it was building Debug/x64 instead.